Fix performance regression
The msgpack map format consists of the map length followed by the map
elements.  Commit f815192 changed the dict serializer to not provide the
number of elements in Serializer::serialize_map calls. This results in a
significant slowdown because, if the length is not known, rmp-serde
serializes the elements in a temporary buffer and writes the
serialization in the output buffer in SerializeMap::end, once the length
is known.

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>