Improve datetime tests
Starting with tzdb-2022b, the offset of the Europe/Amsterdam timezone
for dates up to 1937 Jul 1 is 00:00, not 0:19:32. Because of this,
there is no test coverage for timezones that differ from UTC by a
non-integral number of minutes, and also for the handling of
ormsgpack.OPT_UTC_Z with aware datetime objects.

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>