Remove dependency on RMP
This commit adds and takes into use our own implementation to read and
write the MessagePack formats. The motivation for this change is that
RMP seems to be unmaintained and depends on the unmaintained paste
crate, and that the needed logic is simple and small.
The implementation uses the marker type from RMP.

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>