Fix Dependabot updates of python dependencies
If the project version is declared as dynamic, "uv lock" cannot use
the project metadata and generates the wheel metadata with the build
backend. This causes Dependabot to fail when updating the uv lockfile,
because it invokes uv in a temporary directory containing only the
pyproject.toml file. Work around the issue by specifying the project
version statically. See

https://github.com/astral-sh/uv/blob/0.9.4/crates/uv-pypi-types/src/metadata/metadata_resolver.rs#L175
https://github.com/astral-sh/uv/blob/0.9.4/crates/uv-distribution/src/source/mod.rs#L256
https://github.com/dependabot/dependabot-core/blob/v0.341.0/uv/lib/dependabot/uv/file_updater/lock_file_updater.rb#L201

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>