Optimize serialization
This commit adds a trait to write an array of byte slices, like
write_vectored without the IoSlice wrapper, and implements it for the
serializer writer. Currently, the serializer writes the format byte,
the length and the value of a MessagePack object in separate write_all
calls. A write_all call checks if the PyBytesObject buffer has
sufficient capacity and resizes it if not. While the check is cheap
when there is enough capacity, as the resizing is behind a call to a
cold function, it still adds some overhead. The new trait enables the
serializer to combine the writes and the associated capacity checks.
For example, on macOS arm64 the current code to write a float64 is
compiled to

	ldp x9, x21, [x0]
	add x19, x21, #1
	mov x8, x21
	cmp x19, x9
	b.hi LBB107_3
LBB107_1:
	ldr x9, [x0, #16]
	add x8, x9, x8
	mov w9, #203
	strb w9, [x8, #32]
	str x19, [x0, #8]
	add x20, x21, #9
	ldr x8, [x0]
	cmp x20, x8
	b.hi LBB107_4
LBB107_2:
	fmov x8, d0
	rev x8, x8
	ldr x9, [x0, #16]
	add x9, x9, x19
	str x8, [x9, #32]
	str x20, [x0, #8]
	ldp x29, x30, [sp, #48]
	ldp x20, x19, [sp, #32]
	ldp x22, x21, [sp, #16]
	ldp d9, d8, [sp], #64
	ret
LBB107_3:
	mov x20, x0
	mov x1, x19
	fmov d8, d0
	bl ormsgpack::serialize::writer::BytesWriter::grow
	fmov d0, d8
	mov x0, x20
	ldr x8, [x20, #8]
	b LBB107_1
LBB107_4:
	mov x19, x0
	mov x1, x20
	fmov d8, d0
	bl ormsgpack::serialize::writer::BytesWriter::grow
	fmov d0, d8
	mov x0, x19
	ldr x19, [x19, #8]
	b LBB107_2

while the new code is compiled to

	ldp x9, x8, [x0]
	add x19, x8, #9
	cmp x19, x9
	b.hi LBB107_2
LBB107_1:
	fmov x9, d0
	ldr x10, [x0, #16]
	add x8, x10, x8
	mov w10, #203
	strb w10, [x8, #32]
	rev x9, x9
	stur x9, [x8, #33]
	str x19, [x0, #8]
	ldp x29, x30, [sp, #32]
	ldp x20, x19, [sp, #16]
	ldp d9, d8, [sp], #48
	ret
LBB107_2:
	mov x20, x0
	mov x1, x19
	fmov d8, d0
	bl ormsgpack::serialize::writer::BytesWriter::grow
	fmov d0, d8
	mov x0, x20
	ldr x8, [x20, #8]
	b LBB107_1

Signed-off-by: Emanuele Giaquinta <emanuele.giaquinta@gmail.com>